Refactor + Commit Window + Version Folder
Refactored code to display windows in different positions. Versions folder is now displayed on the right, and headers are now displayed at the bottom. Commit window was modified away from the idealistic graph view to a more simplistic, but easily usable and functional, view. Metadata for files now has added where it's versions are stored, and what the current, latest, version of the file is in the linear timeline