Flipped Predicate, Sharing, Admin Page, and Updated Address Generation
Predicate with which shsared files were stored was flipped for storage and lookup efficiency. Sharing files was fixed to reflect this. Address generation and refresh can now be performed in the hierarchy with a right-click for convenience. The Admin Page was updated to add the UI elements required for the final demonstration scenarios. Work continued on implementing these demonstration scenarios, including a basic SPARQL query testing as well as basic versioned address history retrieval